pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-business-analysis"
version = "0.1.0"
description = "Production-oriented backend and frontend scaffold for business analysis fine-tuning workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "openai>=1.30.0,<2.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pydantic-settings>=2.2.1,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.0.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["backend/src"]
testpaths = ["backend/tests"]

[tool.setuptools.packages.find]
where = ["backend/src"]

